pyproject.toml
[tool.poetry]
name = "whisper-gpt-lineapp-server"
version = "0.1.0"
description = ""
authors = ["satoooh <bighope.lumiere0916@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
line-bot-sdk = "^2.4.2"
requests = "^2.28.2"
openai = "^0.27.2"
mangum = "^0.17.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
